requirements.txt
anthropic==0.84.0
pydantic==2.12.5
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
datasets==3.3.2
spacy==3.7.6
scispacy==0.5.5
# scispacy model: pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz
